+++ test/clj_kondo/types_test.clj
@@ -577,7 +577,15 @@
   (assert-submaps
    '({:file "<stdin>", :row 1, :col 6, :level :error, :message "Expected: number, received: symbol or keyword."})
    (lint! "(inc (or :foo 'bar))"
-          {:linters {:type-mismatch {:level :error}}})))
+          {:linters {:type-mismatch {:level :error}}}))
+  (assert-submaps2
+   [{:file "<stdin>",
+     :row 1,
+     :col 6,
+     :level :error,
+     :message "Expected: number, received: nil."}]
+   (lint! "(inc (or))" {:linters {:type-mismatch {:level :error}}})
+   ))
 
 (deftest cond-test
   (assert-submaps